backend_new/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-filter>=23.5
Pillow>=10.0.0
celery>=5.3.0
redis>=5.0.0
python-decouple>=3.8
drf-spectacular>=0.27.0
